Move isShortenedUrl to its own module
