+++ root/static/scripts/edit/externalLinks.js
@@ -33,6 +33,7 @@ import {compareDatePeriods} from '../common/utility/compareDates.js';
 import {isMalware} from '../url/utility/isGreyedOut.js';
 
 import isPositiveInteger from './utility/isPositiveInteger.js';
+import isShortenedUrl from './utility/isShortenedUrl.js';
 import HelpIcon from './components/HelpIcon.js';
 import RemoveButton from './components/RemoveButton.js';
 import URLInputPopover from './components/URLInputPopover.js';
@@ -607,7 +608,7 @@ export class ExternalLinksEditor
                     because it is known to host malware.`),
         target: URLCleanup.ERROR_TARGETS.URL,
       };
-    } else if (isNewOrChangedLink && isShortened(link.url)) {
+    } else if (isNewOrChangedLink && isShortenedUrl(link.url)) {
       error = {
         message: l(`Please don’t enter bundled/shortened URLs,
                     enter the destination URL(s) instead.`),
@@ -1586,113 +1587,6 @@ function isValidURL(url: string) {
   }
 
   return true;
-}
-
-// For shortener pages which should still be allowed as a host-only link
-const SHORTENER_ALLOWED_HOSTS = [
-  'bruit.app',
-  'distrokid.com',
-  'trac.co',
-];
-
-const URL_SHORTENERS = [
-  'adf.ly',
-  'album.link',
-  'ampl.ink',
-  'amu.se',
-  'artist.link',
-  'band.link',
-  'bfan.link',
-  'biglink.to',
-  'bio.link',
-  'bit.ly',
-  'bitly.com',
-  'backl.ink',
-  'bruit.app',
-  'bstlnk.to',
-  'cli.gs',
-  'deck.ly',
-  'distrokid.com',
-  'ditto.fm',
-  'eventlink.to',
-  'fanlink.to',
-  'ffm.to',
-  'found.ee',
-  'fty.li',
-  'fur.ly',
-  'g.co',
-  'gate.fm',
-  'geni.us',
-  'goo.gl',
-  'hypeddit.com',
-  'hypel.ink',
-  'hyperfollow.com',
-  'hyperurl.co',
-  'is.gd',
-  'kl.am',
-  'laburbain.com',
-  'li.sten.to',
-  'linkco.re',
-  'lnkfi.re',
-  'linkfly.to',
-  'linktr.ee',
-  'listen.lt',
-  'lnk.bio',
-  'lnk.co',
-  'lnk.site',
-  'lnk.to',
-  'lsnto.me',
-  'many.link',
-  'mcaf.ee',
-  'mez.ink',
-  'moourl.com',
-  'music.indiefy.net',
-  'musics.link',
-  'mylink.page',
-  'myurls.bio',
-  'odesli.co',
-  'orcd.co',
-  'owl.ly',
-  'page.link',
-  'pandora.app.link',
-  'podlink.to',
-  'pods.link',
-  'push.fm',
-  'rb.gy',
-  'rubyurl.com',
-  'share.amuse.io',
-  'smarturl.it',
-  'snd.click',
-  'song.link',
-  'songwhip.com',
-  'spinnup.link',
-  'spoti.fi',
-  'sptfy.com',
-  'spread.link',
-  'streamerlinks.com',
-  'streamlink.to',
-  'su.pr',
-  't.co',
-  'tiny.cc',
-  'tinyurl.com',
-  'tourlink.to',
-  'trac.co',
-  'u.nu',
-  'unitedmasters.com',
-  'untd.io',
-  'vyd.co',
-  'yep.it',
-].map(shortener => new RegExp(
-  '^https?://([^/]+\\.)?' +
-  shortener +
-  (SHORTENER_ALLOWED_HOSTS.includes(shortener) ? '/.+' : ''),
-  'i',
-));
-
-function isShortened(url: string) {
-  return URL_SHORTENERS.some(function (shortenerRegex) {
-    return url.match(shortenerRegex) !== null;
-  });
 }
 
 function isGoogleAmp(url: string) {
